Little cleanup

+++ KM_test/src/com/husten/knightmare/graphicalObjects/gasset.java
@@ -20,13 +20,13 @@ public abstract class gasset implements StringConstants {
 
 	public abstract void draw2();
 
-	public abstract float getX();
+	public abstract double getX();
 
-	public abstract float getY();
+	public abstract double getY();
 
-	public abstract void setX(float x);
+	public abstract void setX(double x);
 
-	public abstract void setY(float y);
+	public abstract void setY(double y);
 
 	public abstract Pos getPos();
